requirements.txt
pygame
python-dotenv
pyaudio
sounddevice
websockets
numpy
mss
pyautogui
matplotlib
Pillow
replicate
openai